+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Current SQL statement background highlighting in query editor
+
 ### Fixed
 
 - MongoDB Atlas connections failing to authenticate (#438)